refactor: move SendCommand trait to cc package
