+++ packages/host/src/ZWaveHost.ts
@@ -1,16 +1,3 @@
-import type {
-	CCId,
-	ControllerLogger,
-	SendCommandOptions,
-	SendCommandReturnType,
-} from "@zwave-js/core";
-
-/** Allows sending commands to one or more nodes */
-export interface SendCommand {
-	sendCommand<TResponse extends CCId | undefined = undefined>(
-		command: CCId,
-		options?: SendCommandOptions,
-	): Promise<SendCommandReturnType<TResponse>>;
-}
+import type { ControllerLogger } from "@zwave-js/core";
 
 export type LogNode = Pick<ControllerLogger, "logNode">;
